refactor:> util패키지 local에서 분리 및 Converter.kt 내 메소드 top-level로 변경 #9
